refactor(SimuNEX): add Loads namespace.
