+++ UnityProject/Assets/Testing/PlayMode/Core/Loads/FinTests/FinTests.Derived.cs
@@ -1,4 +1,5 @@
 using SimuNEX;
+using SimuNEX.Loads;
 using UnityEngine;
 
 namespace LoadTests.Fins
