+++ UnityProject/Assets/Testing/PlayMode/Core/Loads/FinTests/FinTests.cs
@@ -1,6 +1,6 @@
 using FactoryTests;
 using NUnit.Framework;
-using SimuNEX;
+using SimuNEX.Loads;
 using SimuNEX.Mechanical;
 using UnityEngine;
