+++ UnityProject/Assets/Testing/PlayMode/Core/Loads/PropellerTests/PropellerTests.Derived.cs
@@ -1,4 +1,5 @@
 using SimuNEX;
+using SimuNEX.Loads;
 using UnityEngine;
 
 namespace LoadTests.Propellers
